+++ scoop/lib/screens/PhotoAddScreen.dart
@@ -1,0 +1,131 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+
+class PhotoAddScreen extends StatefulWidget {
+  const PhotoAddScreen({Key? key}) : super(key: key);
+
+  @override
+  _PhotoAddState createState() => _PhotoAddState();
+}
+
+class _PhotoAddState extends State<PhotoAddScreen>{
+  final formKey = GlobalKey<FormState>();
+
+  String title = '';
+  String link = '';
+  String date = DateFormat.yMd().format(DateTime.now()).toString();
+
+  CollectionReference photo = FirebaseFirestore.instance.collection('photo');
+
+  Future<void> addPhoto() {
+    return photo.add({
+      'title': title,
+      'link': link,
+      'date': date,
+    }).then((value) => print("업로드 성공"))
+    .catchError((error) => print("문제가 발생했습니다: $error"));
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: AppBar(
+        iconTheme: const IconThemeData(
+          color: Colors.black,
+        ),
+        title: const Text(
+          "사진 추가",
+          style: TextStyle(color: Colors.black, fontSize: 17, fontWeight: FontWeight.w700)),
+        backgroundColor: Colors.white,
+        elevation: 0,
+      ),
+      body: Form(
+        key: formKey,
+        child: Column(
+          children: [
+            renderTextFormField(
+              label: '제목', 
+              onSaved: (val) {
+                setState(() {
+                  title = val;
+                });
+              }, 
+              validator: (val) {
+                if (val.length < 1) {
+                  return '제목을 입력해주세요';
+                }
+
+                return null;
+              },
+            ),
+            renderTextFormField(
+              label: '링크', 
+              onSaved: (val) {
+                setState(() {
+                  link = val;
+                });
+              }, 
+              validator: (val) {
+                if (val.length < 1) {
+                  return '링크를 입력해주세요';
+                }
+
+                return null;
+              },
+            ),
+            ElevatedButton(
+              onPressed: () async {
+                if (formKey.currentState!.validate()) {
+                  formKey.currentState!.save();
+                  addPhoto();
+                  Navigator.pop(context);
+                }
+              },
+              style: ButtonStyle(backgroundColor: MaterialStateProperty.all<Color>(Colors.lightBlueAccent)),
+              child: const Text(
+                '업로드',
+                style: TextStyle(
+                  color: Colors.white,
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+renderTextFormField({
+    required String label,
+    required FormFieldSetter onSaved,
+    required FormFieldValidator validator,
+  }) {
+
+    return 
+    Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: Column(
+      children: [
+        Row(
+          children: [
+            Text(
+              label,
+              style: const TextStyle(
+                fontSize: 12.0,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ],
+        ),
+        TextFormField(
+          onSaved: onSaved,
+          validator: validator,
+        ),
+        Container(height: 16.0),
+      ],
+    )
+  );
+  }
